Sentinel-Research/requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
pypdf
unstructured
sentence-transformers
qdrant-client
pinecone-client
python-dotenv
watchdog
fastapi
uvicorn
celery
redis
requests
pytest
langgraph
langchain-huggingface
langchain-qdrant
tf-keras
langchain-ollama
duckduckgo-search
aiofiles

